DMP-2244 pulled up additional methods
