+++ src/main/java/uk/gov/hmcts/darts/arm/service/impl/UnstructuredToArmBatchProcessorSingleElement.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.darts.arm.service.impl;
 
+import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -8,6 +9,8 @@ import org.springframework.stereotype.Service;
 public class UnstructuredToArmBatchProcessorSingleElement {
 
 
+    @Transactional
+    public void insertArmRecord(Integer eodId, String manifestFileName) {
 
-
+    }
 }
